+++ essential-ui/src/routes.tsx
@@ -4,7 +4,7 @@ import NotFoundPage from "./showcase/helpers/NotFound";
 import RootLayout from "./showcase/helpers/RootLayout";
 import ShowcaseLayout from "./showcase/helpers/ShowcaseLayout";
 import HamburgerMenuShowcase from "./showcase/showcases/HamburgerMenu";
-import InputShowcase from "./showcase/showcases/Input";
+import InputShowcase from "./showcase/showcases/Input222";
 
 import { Route } from "react-router-dom";
